backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
redis
pydantic-settings
python-jose[cryptography]
httpx
python-multipart
alembic
